Remove JMock from SequenceBarrierTest
