+++ src/test/java/com/lmax/disruptor/support/DummyEventProcessor.java
@@ -2,15 +2,28 @@ package com.lmax.disruptor.support;
 
 import com.lmax.disruptor.EventProcessor;
 import com.lmax.disruptor.Sequence;
+import com.lmax.disruptor.SingleProducerSequencer;
+
+import java.util.concurrent.atomic.AtomicBoolean;
 
 public class DummyEventProcessor implements EventProcessor
 {
     private final Sequence sequence;
-    private boolean isRunning;
+    private final AtomicBoolean running = new AtomicBoolean(false);
 
     public DummyEventProcessor(Sequence sequence)
     {
         this.sequence = sequence;
+    }
+
+    public DummyEventProcessor()
+    {
+        this(new Sequence(SingleProducerSequencer.INITIAL_CURSOR_VALUE));
+    }
+
+    public void setSequence(long sequence)
+    {
+        this.sequence.set(sequence);
     }
 
     @Override
@@ -22,18 +35,21 @@ public class DummyEventProcessor implements EventProcessor
     @Override
     public void halt()
     {
-        isRunning = false;
+        running.set(false);
     }
 
     @Override
     public boolean isRunning()
     {
-        return isRunning;
+        return running.get();
     }
 
     @Override
     public void run()
     {
-        isRunning = true;
+        if (!running.compareAndSet(false, true))
+        {
+            throw new IllegalStateException("Already running");
+        }
     }
 }
